% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateDF.R
\name{CalculateDF}
\alias{CalculateDF}
\title{Calculate the difference of metabolic subpathway activity}
\usage{
CalculateDF(spwmatrix, smaple_class, casename = "", controlname = "")
}
\arguments{
\item{spwmatrix}{Metabolic subpathway activity matrix (the result of function `SubpathwayMatrix`),}

\item{smaple_class}{Sample phenotype vector in metabolic subpathway activity matrix.}

\item{casename}{Disease sample phenotype label.}

\item{controlname}{Control sample phenotype label.}
}
\description{
For the metabolic subpathway activity matrix, this method can calculate the difference and significance of the metabolic subpathway activity between disease and normal.
}
\details{
CalculateDF
}
\examples{
library(limma)
# Get the metabolic subpathway matrix.
Spwmatrix<-get("Spwmatrix")
spwDF<-CalculateDF(Spwmatrix,colnames(Spwmatrix),"cancer","control")
}
