% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFspw.R
\docType{data}
\name{DFspw}
\alias{DFspw}
\title{Difference analysis results of metabolic subpathways}
\format{
A list.
}
\usage{
DFspw
}
\description{
The difference analysis result of teh function CalculateDF is for the metabolic subpathway matrix.
}
\examples{
\donttest{data(DFspw)}
}
\keyword{datasets}
