% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geneexp.R
\docType{data}
\name{Geneexp}
\alias{Geneexp}
\title{Simulated gene expression data}
\format{
A matrix with 3000 genes and 40 samples.
}
\usage{
Geneexp
}
\description{
Simulated normalized gene expression profile data.
}
\examples{
\donttest{data(Geneexp)}
}
\keyword{datasets}
