% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPackageLoaded.R
\name{isPackageLoaded}
\alias{isPackageLoaded}
\title{crmSubpathway internal function}
\usage{
isPackageLoaded(name)
}
\arguments{
\item{name}{A string. The name of the R package which determines whether it is loaded.}
}
\value{
A string, TRUE or FALSE.
}
\description{
Determine if the package is loaded. If the package is not loaded,
the program will prompt the user.
}
\details{
isPackageLoaded
}
\examples{
\donttest{isPackageLoaded("pheatmap")}
}
