% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_clique.R
\name{k_clique}
\alias{k_clique}
\title{k-clique algorithm}
\usage{
k_clique(file.path = "", file.names = "", K = 4)
}
\arguments{
\item{file.path}{Path of KGML files (XML format).}

\item{file.names}{A character vector of names of KGML files.}

\item{K}{An integer. A distance similarity parameter (default: 4).}
}
\description{
Mining metabolic subpathways through k-clique algorithm.
}
\details{
k_clique


In this method, we convert the each metabolic pathways in the KEGG database into an undirected network diagram by connecting
two genes (enzymes) into an edge if there is a common compound in their corresponding reactions. Based on the distance
similarity between genes, we use the K-clique method in social network analysis to mine subpathways and the distance between
all genes is no exceeding K is determined as the metabolic subpathway.At the same time, our method can also be applied to other
non-metabolic pathways of KEGG database.
}
\examples{
library(graph);
library(RBGL);
library(igraph);
library(XML);
file.path<-paste(system.file(package="crmSubpathway"),"/extdata/",sep="")
file.names<-c("hsa00010.xml","hsa00020.xml")
spwlist<-k_clique(file.path,file.names)
}
