% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metspwlist.R
\docType{data}
\name{Metspwlist}
\alias{Metspwlist}
\title{Metabolic subpathway list data}
\format{
A list.
}
\usage{
Metspwlist
}
\description{
List of metabolic sub-pathways excavated from the KEGG database.
}
\examples{
\donttest{data(Metspwlist)}
}
\keyword{datasets}
