% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspwheatmap.R
\name{plotspwheatmap}
\alias{plotspwheatmap}
\title{Plot heat map of the metabolic subpathway activity matrix}
\usage{
plotspwheatmap(
  DF_result,
  PValue_threshold = 0.01,
  main = "",
  show.rownames = TRUE,
  show.colnames = TRUE,
  fontsize = 10,
  fontsize.row = 10,
  fontsize.col = 10,
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  scale = "none"
)
}
\arguments{
\item{DF_result}{A list of result data generated by function `CalculateDF()`.}

\item{PValue_threshold}{Threshold for screening significant metabolic subpathways.}

\item{main}{an overall title for the heatmap.}

\item{show.rownames}{Boolean specifying if row names are be shown.}

\item{show.colnames}{Boolean specifying if column names are be shown.}

\item{fontsize}{Base fontsize for the plot (default: 10).}

\item{fontsize.row}{Fontsize for rownames (default: 10).}

\item{fontsize.col}{Fontsize for colnames (default: 10).}

\item{cluster_rows}{boolean values determining if rows should be clustered or hclust object.}

\item{cluster_cols}{boolean values determining if columns should be clustered or hclust object.}

\item{scale}{Character indicating if the values should be centered}
}
\value{
A heat map.
##' @examples
require(pheatmap)
DFspw<-get("DFspw")
plotspwheatmap(DFspw,cluster_rows = TRUE,show.colnames = FALSE)
}
\description{
According to the threshold of the set P value, The `plotspwheatmap()` function plots a heat map of the significant metabolic subpathways.
}
\details{
plotspwheatmap
}
