% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redup.R
\name{redup}
\alias{redup}
\title{Remove redundant metabolic subpathways}
\usage{
redup(spw)
}
\arguments{
\item{spw}{list of metabolic subpathway}
}
\value{
A list.
}
\description{
Determine if the package is loaded. If the package is not loaded,
the program will prompt the user.
}
\details{
redup
}
