% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spwmatrix.R
\docType{data}
\name{Spwmatrix}
\alias{Spwmatrix}
\title{Metabolism subpathway activity matrix}
\format{
A matrix.
}
\usage{
Spwmatrix
}
\description{
Metabolic subpathway activity matrix for case demonstration.
}
\examples{
\donttest{data(Spwmatrix)}
}
\keyword{datasets}
