% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpwNetworkData.R
\docType{data}
\name{SpwNetworkData}
\alias{SpwNetworkData}
\title{Subpathway network structure data}
\format{
A list containing 1598 subpathway network.
}
\usage{
SpwNetworkData
}
\description{
A list to store the network data of the genes contained in the subpathway.
}
\examples{
\donttest{data(SpwNetworkData)}
}
\keyword{datasets}
