% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SubpathwayMatrix.R
\name{SubpathwayMatrix}
\alias{SubpathwayMatrix}
\title{Get metabolic subpathway activity matrix}
\usage{
SubpathwayMatrix(
  Gexp,
  Subpathway.list,
  method = "gsva",
  kcdf = "Gaussian",
  min.sz = 1,
  max.sz = Inf,
  parallel.sz = 0
)
}
\arguments{
\item{Gexp}{Matrix of gene expression values (rows are genes, columns are samples).}

\item{Subpathway.list}{The subpathway list data. We have stored the processed human metabolic subpathway data set in variable `Metspwlist`.
Users can also set their own data set of subpathways.}

\item{method}{Method to employ in the estimation of subpathway
enrichment scores per sample. By default this is set to `gsva` (Hänzelmann
et al, 2013) and other options are `ssgsea` (Barbie et al, 2009).}

\item{kcdf}{Character string denoting the kernel to use during the
non-parametric estimation of the cumulative distribution function of
expression levels across samples when `spw.score.method="gsva"`. By default,
`kcdf="Gaussian"` which is suitable when input expression values are
continuous, such as microarray fluorescent units in logarithmic scale,
RNA-seq log-CPMs, log-RPKMs or log-TPMs. When input expression values are
integer counts, such as those derived from RNA-seq experiments, then this
argument should be set to `kcdf="Poisson"`.}

\item{min.sz}{Removes subpathways that contain fewer genes than `spw.min.sz` (default: 10).}

\item{max.sz}{Removes subpathways that contain more genes than `spw.max.sz` (default: Inf).}

\item{parallel.sz}{Number of processors to use when doing the calculations in
parallel (default value: 1). If parallel.sz=0,
then it will use all available core processors unless we set this argument
with a smaller number.}
}
\description{
Get metabolic subpathway activity matrix.
}
\details{
SubpathwayMatrix


Our method assesses the relative enrichment of metabolic subpathways across samples using a non-parametric approach.
Conceptually, this method transforms a p-gene by n-sample gene expression matrix into a g-subpathway by n-sample metabolic subpathway
enrichment matrix.
}
\examples{
library(GSVA)
library(parallel)
# Get the dataset of metabolic subpathways we have processed.
Metspwlist<-get("Metspwlist")
# Get the gene expression profile of the case.
Geneexp<-get("Geneexp")
Spwmatrix<-SubpathwayMatrix(Geneexp,Metspwlist)
}
