% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_clique.R
\name{k_clique}
\alias{k_clique}
\title{k_clique algorithm}
\usage{
k_clique(file.path = "", file.names = "", K = 4)
}
\arguments{
\item{file.path}{Path of KGML files (XML format).}

\item{file.names}{A character vector of names of KGML files.}

\item{K}{An integer. A distance similarity parameter.}
}
\description{
Mining metabolic  subpathways through k_clique algorithm.
}
\details{
k_clique
}
\examples{
library(graph);
library(RBGL);
library(igraph);
library(XML);
file.path<-paste(system.file(package="crmSubpathway"),"/extdata/",sep="")
file.names<-c("hsa00010.xml","hsa00020.xml")
spwlist<-k_clique(file.path,file.names)
}
